docai_async_reference_architecture/auto_deploy_v8/CFScript/requirements.txt
functions-framework==3.*
google-cloud-documentai>=2.7.0
pandas>=1.5.2
google-cloud-storage>=2.7.0
google-cloud-firestore>=2.8.0
firebase-admin>=6.0.1
